requirement.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
pydantic==1.10.11
python-jose[cryptography]
passlib[bcrypt]
PyJWT==2.8.0
python-dateutil==2.8.2
aiofiles==23.1.0
alembic==1.13.0
python-dotenv==1.0.0
python-multipart==0.0.20
